created a seperate view for going back to origin, removed score from text replies and photo replies, added missing file for reventing voting limit, refactoring unseen activity.
