+++ unconnectedreddit/templates/reply.html
@@ -79,13 +79,6 @@
                 {% endif %}
             {% endif %}
            <a href="{% url 'profile' slug=reply.submitted_by__username %}" class="clb"><bdi>{% if reply.submitted_by__username in ensured %}<img src="{{ STATIC_URL }}img/pstar.svg" alt="*" height="11" width="11">&nbsp;{% endif %}<b>{{ reply.submitted_by__username }}</b></bdi></a>
-            <span class="cxs sp">
-                {% if reply.submitted_by__userprofile__score < 0 %}
-                    <span class="cg">(1)</span>
-                {% else %}
-                    <span class="cg">({{ reply.submitted_by__userprofile__score }})</span>
-                {% endif %}
-            </span>
                :&nbsp;<span class="clb lsp bw"><bdi>{% emoticons %}{{ reply.description }}{% endemoticons %}</bdi></span><span class="mls mrl cxs cgy sp"><bdi>{{ reply.submitted_on|human_time }}</bdi></span>{% if user_id == parent_submitter_id %}<form action="{% url 'hide_jawab' publicreply_id=reply.id link_id=parent.pk %}" style="display:inline" method="POST" class="mrl">{% csrf_token %}<button type="submit" class="pg_ok_res" style="background:0 0;border:none;padding:0"><img src="{{ STATIC_URL }}img/report_dark.svg" width="16" height="16" style="vertical-align:middle"></button></form>{% endif %}<span class="mrs">{% block_button reply.submitted_by reply.submitted_by__username user_id parent.pk '9' STATIC_URL %}</span>{% group_button user_id reply.submitted_by parent.pk 'home' '9' reply.submitted_by__username %}
         {% endif %}
         </div>
